address @austinabell comments
