+++ types/account.go
@@ -18,12 +18,6 @@ import (
 var _ exported.Account = (*Account)(nil)
 var _ exported.GenesisAccount = (*Account)(nil)
 
-const (
-	// DenomDefault defines the single coin type/denomination supported in
-	// Ethermint.
-	DenomDefault = "photon"
-)
-
 // ----------------------------------------------------------------------------
 // Main Ethermint account
 // ----------------------------------------------------------------------------
@@ -33,7 +27,7 @@ func init() {
 }
 
 // Account implements the auth.Account interface and embeds an
-// auth.BaseAccount type. It is compatible with the auth.AccountMapper.
+// auth.BaseAccount type. It is compatible with the auth.AccountKeeper.
 type Account struct {
 	*auth.BaseAccount
 
@@ -45,9 +39,9 @@ type Account struct {
 	CodeHash []byte
 }
 
-// ProtoBaseAccount defines the prototype function for BaseAccount used for an
-// account mapper.
-func ProtoBaseAccount() exported.Account {
+// ProtoAccount defines the prototype function for BaseAccount used for an
+// AccountKeeper.
+func ProtoAccount() exported.Account {
 	return &Account{
 		BaseAccount: &auth.BaseAccount{},
 		CodeHash:    ethcrypto.Keccak256(nil),
@@ -75,39 +69,6 @@ func (acc Account) SetBalance(amt sdk.Int) {
 	if err := acc.SetCoins(coins); err != nil {
 		panic(fmt.Sprintf("Could not set coins for address %s", acc.GetAddress()))
 	}
-}
-
-// ----------------------------------------------------------------------------
-// Code & Storage
-// ----------------------------------------------------------------------------
-
-type (
-	// Code is account Code type alias
-	Code []byte
-	// Storage is account storage type alias
-	Storage map[ethcmn.Hash]ethcmn.Hash
-)
-
-func (c Code) String() string {
-	return string(c)
-}
-
-func (c Storage) String() (str string) {
-	for key, value := range c {
-		str += fmt.Sprintf("%X : %X\n", key, value)
-	}
-
-	return
-}
-
-// Copy returns a copy of storage.
-func (c Storage) Copy() Storage {
-	cpy := make(Storage)
-	for key, value := range c {
-		cpy[key] = value
-	}
-
-	return cpy
 }
 
 type ethermintAccountPretty struct {
