+++ types/params.go
@@ -5,4 +5,8 @@ const (
 	DefaultGasPrice = 20
 	// DefaultRPCGasLimit is default gas limit for RPC call operations
 	DefaultRPCGasLimit = 10000000
+
+	// DenomDefault defines the single coin type/denomination supported in
+	// Ethermint.
+	DenomDefault = "photon"
 )
