+++ x/evm/types/state_object.go
@@ -7,7 +7,6 @@ import (
 	"math/big"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/cosmos/cosmos-sdk/x/auth"
 	authexported "github.com/cosmos/cosmos-sdk/x/auth/exported"
 
 	"github.com/cosmos/ethermint/types"
@@ -79,21 +78,9 @@ type (
 )
 
 func newStateObject(db *CommitStateDB, accProto authexported.Account) *stateObject {
-	var (
-		ethermintAccount *types.Account
-		ok               bool
-	)
-	ethermintAccount, ok = accProto.(*types.Account)
+	ethermintAccount, ok := accProto.(*types.Account)
 	if !ok {
-		// check if account is a BaseAccount
-		baseAcc, ok := accProto.(*auth.BaseAccount)
-		if !ok {
-			panic(fmt.Sprintf("invalid account type for state object: %T", accProto))
-		}
-		ethermintAccount = &types.Account{
-			BaseAccount: baseAcc,
-			CodeHash:    emptyCodeHash,
-		}
+		panic(fmt.Sprintf("invalid account type for state object: %T", accProto))
 	}
 
 	// set empty code hash
